# optimize rust code